Separate libpostal core initialization, fix prev. commit test fail
In this commit, I believe I have fixed the "root cause" of the problem
that ets@ commented about on the first version of my pull request to
the JPostal repo, https://github.com/openvenues/jpostal/pull/46.

In my initial commit/PR, I had misinterpreted how the builders worked,
because it's a bit counterintuitive. I had naively thought that their
`setDefaultOptions` methods were pushing global/static configuration
into libpostal and thus only had to run once, whereas the truth is they
*pull* global/static configuration from libpostal.

I looked at different ways to solve this. The ideal would be to
refactor the builders to contain boxed objects where `null` represents
the default value, and have the JNI C code check each option field and
if its null just pull the value from the corresponding libpostal global
default field at the time that parse/expand is called. However, I
rejected this option because it would result in a backwardly-
incompatible change to the builders/options.

In the end, this commit is basically putting the `setDefaultOptions`
logic back the way it was before my previous commit. However, there
is one improvement, which is that I have factored the native library
load into a single location (driven by the package-internal singleton
class `LibPostal`). This is really logically necessary because the
expander and parser both rely on the same underlying library setup
and they should not try to set it up in conflicting ways.

My improvement also allows core use of thread-unsafe global library
constructs to be used in a thread-safe manner within Java by making
it synchronized on the singleton `LibPostal` instance rather than
on the `AddressParser` and `AddressExpander` classes, since there
are two of those classes, which means having some global init
synchronized on one, and some on the other, can lead to race
conditions.